+++ lib/rubocop/cop/style/single_space_before_first_arg.rb
@@ -6,12 +6,18 @@ module RuboCop
       # Checks that exactly one space is used between a method name and the
       # first argument for method calls without parentheses.
       #
-      # @example
+      # Alternatively, extra spaces can be added to align the argument with
+      # something on a preceding or following line, if the AllowForAlignment
+      # config parameter is true.
       #
+      # @example
+      #   @bad
       #   something  x
       #   something   y, z
       #
       class SingleSpaceBeforeFirstArg < Cop
+        include PrecedingFollowingAlignment
+
         MSG = 'Put one space between the method name and the first argument.'
 
         def on_send(node)
@@ -24,6 +30,7 @@ module RuboCop
 
           arg1 = args.first.loc.expression
           return if arg1.line > node.loc.line
+          return if allow_for_alignment? && aligned_with_something?(arg1)
 
           arg1_with_space = range_with_surrounding_space(arg1, :left)
           space = Parser::Source::Range.new(arg1.source_buffer,